HX-442 HX-298 HX-245 HX-498 HX-445 HX-583 HX-143

+++ feature-account-impl/src/main/java/io/novafoundation/nova/feature_account_impl/presentation/menu/change_autolock/ChangeAutoLockPresenter.kt
@@ -38,7 +38,7 @@ class ChangeAutoLockPresenter @Inject constructor(
 
     fun onNextClick() {
         presenterScope.launch {
-            val time = newTime.filter { it.isDigit() }.toInt().absoluteValue
+            val time = newTime.filter { it.isDigit() }.toDouble().absoluteValue
             if (time / 60 > 24) {
                 showError(resourceManager.getString(R.string.error_auto_lock))
             } else {
